+++ scripts/winesapos-upgrade.sh
@@ -92,55 +92,55 @@ echo "Upgrading GRUB menu complete."
 
 echo "Upgrading packages..."
 echo "Installing Heroic Games Launcher for Epic Games Store games if needed..."
-sudo -u stick yay --noconfirm -S --needed heroic-games-launcher-bin
+sudo -u winesap yay --noconfirm -S --needed heroic-games-launcher-bin
 echo "Installing 'smartmontools' if needed..."
 pacman -S --needed --noconfirm smartmontools
 echo "Upgrading packages complete."
 
 echo "Upgrading desktop shortcuts..."
-if [ ! -f /home/stick/Desktop/heroic_games_launcher.desktop ]; then
-    cp /usr/share/applications/heroic.desktop /home/stick/Desktop/heroic_games_launcher.desktop
-    sed -i s'/Exec=\/opt\/Heroic\/heroic\ \%U/Exec=\/usr\/bin\/gamemoderun \/opt\/Heroic\/heroic\ \%U/'g /home/stick/Desktop/heroic_games_launcher.desktop
-    crudini --set /home/stick/Desktop/heroic_games_launcher.desktop "Desktop Entry" Name "Heroic Games Launcher - GameMode"
+if [ ! -f /home/winesap/Desktop/heroic_games_launcher.desktop ]; then
+    cp /usr/share/applications/heroic.desktop /home/winesap/Desktop/heroic_games_launcher.desktop
+    sed -i s'/Exec=\/opt\/Heroic\/heroic\ \%U/Exec=\/usr\/bin\/gamemoderun \/opt\/Heroic\/heroic\ \%U/'g /home/winesap/Desktop/heroic_games_launcher.desktop
+    crudini --set /home/winesap/Desktop/heroic_games_launcher.desktop "Desktop Entry" Name "Heroic Games Launcher - GameMode"
 fi
-if [ ! -f /home/stick/Desktop/lutris.desktop ]; then
-    cp /usr/share/applications/net.lutris.Lutris.desktop /home/stick/Desktop/lutris.desktop
-    sed -i s'/Exec=lutris\ \%U/Exec=\/usr\/bin\/gamemoderun \/usr\/bin\/lutris\ \%U/'g /home/stick/Desktop/lutris.desktop
-    crudini --set /home/stick/Desktop/lutris.desktop "Desktop Entry" Name "Lutris - GameMode"
+if [ ! -f /home/winesap/Desktop/lutris.desktop ]; then
+    cp /usr/share/applications/net.lutris.Lutris.desktop /home/winesap/Desktop/lutris.desktop
+    sed -i s'/Exec=lutris\ \%U/Exec=\/usr\/bin\/gamemoderun \/usr\/bin\/lutris\ \%U/'g /home/winesap/Desktop/lutris.desktop
+    crudini --set /home/winesap/Desktop/lutris.desktop "Desktop Entry" Name "Lutris - GameMode"
 fi
-if [ ! -f /home/stick/Desktop/steam_native.desktop ]; then
-    cp /usr/share/applications/steam-native.desktop /home/stick/Desktop/steam_native.desktop
-    sed -i s'/Exec=\/usr\/bin\/steam\-native\ \%U/Exec=\/usr\/bin\/gamemoderun \/usr\/bin\/steam\-native\ \%U/'g /home/stick/Desktop/steam_native.desktop
-    crudini --set /home/stick/Desktop/steam_native.desktop "Desktop Entry" Name "Steam (Native) - GameMode"
+if [ ! -f /home/winesap/Desktop/steam_native.desktop ]; then
+    cp /usr/share/applications/steam-native.desktop /home/winesap/Desktop/steam_native.desktop
+    sed -i s'/Exec=\/usr\/bin\/steam\-native\ \%U/Exec=\/usr\/bin\/gamemoderun \/usr\/bin\/steam\-native\ \%U/'g /home/winesap/Desktop/steam_native.desktop
+    crudini --set /home/winesap/Desktop/steam_native.desktop "Desktop Entry" Name "Steam (Native) - GameMode"
 fi
-if [ ! -f /home/stick/Desktop/steam_runtime.desktop ]; then
-    cp /usr/lib/steam/steam.desktop /home/stick/Desktop/steam_runtime.desktop
-    sed -i s'/Exec=\/usr\/bin\/steam\-runtime\ \%U/Exec=\/usr\/bin\/gamemoderun \/usr\/bin\/steam-runtime\ \%U/'g /home/stick/Desktop/steam_runtime.desktop
-    crudini --set /home/stick/Desktop/steam_runtime.desktop "Desktop Entry" Name "Steam (Runtime) - GameMode"
+if [ ! -f /home/winesap/Desktop/steam_runtime.desktop ]; then
+    cp /usr/lib/steam/steam.desktop /home/winesap/Desktop/steam_runtime.desktop
+    sed -i s'/Exec=\/usr\/bin\/steam\-runtime\ \%U/Exec=\/usr\/bin\/gamemoderun \/usr\/bin\/steam-runtime\ \%U/'g /home/winesap/Desktop/steam_runtime.desktop
+    crudini --set /home/winesap/Desktop/steam_runtime.desktop "Desktop Entry" Name "Steam (Runtime) - GameMode"
 fi
 
 for i in \
   freeoffice-planmaker.desktop \
   freeoffice-presentations.desktop \
   freeoffice-textmaker.desktop
-    do if [ ! -f "/home/stick/Desktop/${i}" ]; then
-        cp "/usr/share/applications/${i}" "/home/stick/Desktop/${i}"
+    do if [ ! -f "/home/winesap/Desktop/${i}" ]; then
+        cp "/usr/share/applications/${i}" "/home/winesap/Desktop/${i}"
     fi
 done
 
-if [ ! -f /home/stick/Desktop/google-chrome.desktop ]; then
-    cp /usr/share/applications/google-chrome.desktop /home/stick/Desktop/
+if [ ! -f /home/winesap/Desktop/google-chrome.desktop ]; then
+    cp /usr/share/applications/google-chrome.desktop /home/winesap/Desktop/
 fi
-if [ ! -f /home/stick/Desktop/qdirstat.desktop ]; then
-    cp /usr/share/applications/qdirstat.desktop /home/stick/Desktop/
+if [ ! -f /home/winesap/Desktop/qdirstat.desktop ]; then
+    cp /usr/share/applications/qdirstat.desktop /home/winesap/Desktop/
 fi
 # Fix permissions on the desktop shortcuts.
-chmod +x /home/stick/Desktop/*.desktop
-chown -R stick: /home/stick/Desktop/*.desktop
+chmod +x /home/winesap/Desktop/*.desktop
+chown -R winesap: /home/winesap/Desktop/*.desktop
 echo "Upgrading desktop shortcuts complete."
 
 echo "Uprading by adding Proton GE..."
-ls -1 /home/stick/.local/share/Steam/compatibilitytools.d/ | grep -v -P ".tar.gz$" | grep -q -P "^Proton.*GE.*"
+ls -1 /home/winesap/.local/share/Steam/compatibilitytools.d/ | grep -v -P ".tar.gz$" | grep -q -P "^Proton.*GE.*"
 if [ $? -eq 0 ]; then
     echo "Proton GE already installed. Skipping."
 else
@@ -149,12 +149,12 @@ else
     chmod +x /usr/local/bin/ge-install-manager
     # The '/tmp/' directory will not work as a 'tmp_path' for 'ge-install-manager' due to a
     # bug relating to calculating storage space on ephemeral file systems. As a workaround,
-    # we use '/home/stick/tmp' as the temporary path.
+    # we use '/home/winesap/tmp' as the temporary path.
     # https://github.com/toazd/ge-install-manager/issues/3
-    mkdir -p /home/stick/tmp/ /home/stick/.config/ge-install-manager/ /home/stick/.steam/root/compatibilitytools.d/
-    cp ../files/ge-install-manager.conf /home/stick/.config/ge-install-manager/
-    chown -R stick: /home/stick/tmp /home/stick/.config /home/stick/.steam
-    sudo -u stick ge-install-manager -i Proton-6.5-GE-2
+    mkdir -p /home/winesap/tmp/ /home/winesap/.config/ge-install-manager/ /home/winesap/.steam/root/compatibilitytools.d/
+    cp ../files/ge-install-manager.conf /home/winesap/.config/ge-install-manager/
+    chown -R winesap: /home/winesap/tmp /home/winesap/.config /home/winesap/.steam
+    sudo -u winesap ge-install-manager -i Proton-6.5-GE-2
 fi
 echo "Uprading by adding Proton GE complete."
 
@@ -163,7 +163,7 @@ if [ -f /usr/bin/protontricks ]; then
     echo "'protontricks' is already installed. Skipping."
 else
     echo "'protontricks' was not found. Installing now..."
-    sudo -u stick yay --noconfirm -S --needed protontricks
+    sudo -u winesap yay --noconfirm -S --needed protontricks
 fi
 echo "Upgrading by adding 'protontricks' program complete."